Move purify method into separate file
