+++ src/lib/rpf_purify.cpp
@@ -1,0 +1,1058 @@
+#include "rpf.hpp"
+
+
+void RandomPlantedForest::purify_3()
+{
+
+  // go through all n_trees families
+  for (auto &curr_family : this->tree_families)
+  {
+
+    // lim_list is a list giving for each variable all interval end-points
+    std::vector<std::vector<double>> lim_list = get_lim_list(curr_family);
+
+    // initialize values and individuals for each tree in family
+    std::vector<grid::NDGrid> grids(curr_family.size() - 1);
+    std::vector<utils::Matrix<int>> individuals(curr_family.size() - 1);
+    std::vector<utils::Matrix<std::vector<double>>> values(curr_family.size() - 1);
+    std::vector<utils::Matrix<std::vector<double>>> values_old(curr_family.size() - 1);
+    std::vector<std::set<int>> variables(curr_family.size() - 1);
+
+    //  ------------- setup finer grid  -------------
+
+    int tree_index = 0;
+    for (const auto &curr_tree : curr_family)
+    {
+
+      if (curr_tree.first == std::set<int>{0})
+      {
+
+        // values[tree_index] = rpf::Matrix<std::vector<double>>(dimensions, std::vector<double>(value_size, 0)); // changed
+        continue; // ignore null tree
+      }
+
+      // fill space with dimensions
+      std::vector<int> dimensions;
+      for (const auto &dim : curr_tree.first)
+      {
+        dimensions.push_back(lim_list[dim - 1].size()); // size - 1 ? WICHTIG
+      }
+
+      // setup grid for leaf indices
+      auto grid = grid::NDGrid(dimensions);
+
+      // initialize data for current tree
+      grids[tree_index] = grid;
+      individuals[tree_index] = utils::Matrix<int>(dimensions, 0);
+      values[tree_index] = utils::Matrix<std::vector<double>>(dimensions, std::vector<double>(value_size, 0));     // changed
+      values_old[tree_index] = utils::Matrix<std::vector<double>>(dimensions, std::vector<double>(value_size, 0)); // changed
+      variables[tree_index] = curr_tree.first;
+
+      // fill grid points with individuals and values
+      while (!grid.nextPoint())
+      {
+
+        std::vector<int> gridPoint = grid.getPoint();
+
+        bool in_leaf = true;
+
+        // go through sample points to sum up individuals
+        for (const auto &feature_vec : X)
+        {
+          int dim_index = 0;
+          in_leaf = true;
+          for (const auto &dim : curr_tree.first)
+          {
+            double val = feature_vec[dim - 1];
+            if (!((val >= lim_list[dim - 1][gridPoint[dim_index]]) && (val < lim_list[dim - 1][gridPoint[dim_index] + 1])))
+              in_leaf = false; //Break here?
+            ++dim_index;
+          }
+
+          // consider individuals only if all in
+          if (in_leaf)
+            individuals[tree_index][gridPoint] += 1;
+        }
+
+        // go through leaves of tree to sum up values
+        for (const auto &leaf : curr_tree.second->get_leaves())
+        {
+
+          in_leaf = true;
+          int dim_index = 0;
+          for (const auto &dim : curr_tree.first)
+          {
+            // consider values only if all in
+            if (!((leaf.intervals[dim - 1].first <= lim_list[dim - 1][gridPoint[dim_index]]) && (leaf.intervals[dim - 1].second >= lim_list[dim - 1][gridPoint[dim_index] + 1])))
+              in_leaf = false; // Break here? 
+            ++dim_index;
+          }
+
+          // sum up values
+          if (in_leaf)
+          {
+
+            values[tree_index][gridPoint] += leaf.value;     // todo: multiclass
+            values_old[tree_index][gridPoint] += leaf.value; // todo: multiclass
+          }
+        }
+      }
+
+      ++tree_index;
+    }
+
+    // Rcout << variables.size();
+    // for(int i = 0; i<variables.size(); ++i){
+    //
+    //   // Rcout << variables[i].size();
+    //
+    //   for(auto dim: variables[i]) Rcout << dim << ",";
+    //
+    //   //  Rcout << variables[i][j] << ",";
+    //   //}
+    //
+    //   Rcout << std::endl;
+    // }
+
+    // ------------- create new trees -------------
+
+    // insert null tree
+    grids.insert(grids.begin(), grid::NDGrid());
+    values.insert(values.begin(), utils::Matrix<std::vector<double>>(std::vector<int>{1}, std::vector<double>(value_size, 0)));
+    values_old.insert(values_old.begin(), utils::Matrix<std::vector<double>>(std::vector<int>{1}, std::vector<double>(value_size, 0)));
+    individuals.insert(individuals.begin(), utils::Matrix<int>(std::vector<int>{1}));
+    variables.insert(variables.begin(), std::set<int>{0});
+
+    // recap maximum number of dimensions of current family
+    unsigned int curr_max = curr_family.rbegin()->first.size();
+
+    while (curr_max > 1)
+    {
+
+      auto keys = getKeys(curr_family);
+      // go through split dimensions of all trees
+      for (std::vector<std::set<int>>::reverse_iterator key = keys.rbegin(); key != keys.rend(); ++key)
+      {
+        auto &curr_tree = curr_family[(*key)];
+        std::set<int> curr_dims = curr_tree->split_dims;
+        // check if number of dims same as current max_interaction
+        if (curr_dims.size() == curr_max)
+        {
+          // go through feature dims
+          int dim_index = 0;
+          for (int feature_dim = 1; feature_dim <= feature_size; ++feature_dim)
+          {
+            // continue only if dim in current tree
+            if (curr_tree->split_dims.count(feature_dim) != 0)
+            {
+              std::set<int> tree_dims = curr_tree->split_dims;
+              tree_dims.erase(tree_dims.find(feature_dim)); // remove current feature dim from current tree
+              // check if tree with dimensions exists, if not create
+              std::shared_ptr<DecisionTree> tree = treeExists(tree_dims, curr_family);
+              if (!tree)
+              {
+                // get index of old and new tree
+                auto old_tree_index = std::distance(std::begin(curr_family), curr_family.find(curr_tree->get_split_dims()));
+                curr_family.insert(std::make_pair(tree_dims, std::make_shared<DecisionTree>(DecisionTree(tree_dims))));
+                auto tree_index = std::distance(std::begin(curr_family), curr_family.find(tree_dims));
+                // remove matrix dimension of respective variable
+                std::vector<int> matrix_dimensions = values[old_tree_index].dims;
+                // std::vector<int> matrix_dimensions = values_old[old_tree_index].dims;
+
+                // Rcout << typeof(matrix_dimensions.begin()) << std::endl;
+
+                matrix_dimensions.erase(matrix_dimensions.begin() + dim_index);
+                // initialize data for new tree
+                auto grid = grid::NDGrid(matrix_dimensions);
+                grids.insert(grids.begin() + tree_index, grid);
+                values.insert(values.begin() + tree_index, utils::Matrix<std::vector<double>>(matrix_dimensions, std::vector<double>(value_size, 0)));
+                values_old.insert(values_old.begin() + tree_index, utils::Matrix<std::vector<double>>(matrix_dimensions, std::vector<double>(value_size, 0)));
+                individuals.insert(individuals.begin() + tree_index, utils::Matrix<int>(matrix_dimensions));
+                variables.insert(variables.begin() + tree_index, tree_dims);
+                // fill individuals of new trees
+                while (!grid.nextPoint())
+                {
+                  std::vector<int> gridPoint = grid.getPoint();
+                  bool in_leaf = true;
+                  // go through sample points to sum up individuals
+                  for (const auto &feature_vec : X)
+                  {
+                    int dim_index2 = 0;
+                    in_leaf = true;
+                    for (const auto &dim : tree_dims)
+                    {
+                      double val = feature_vec[dim - 1];
+                      if (!((val >= lim_list[dim - 1][gridPoint[dim_index2]]) && (val < lim_list[dim - 1][gridPoint[dim_index2] + 1])))
+                        in_leaf = false;
+                      ++dim_index2;
+                    }
+                    // consider individuals only if all in
+                    if (in_leaf)
+                      individuals[tree_index][gridPoint] += 1;
+                  }
+                }
+              }
+              dim_index++;
+            }
+          }
+        }
+      }
+      // update currently considered dimension size
+      --curr_max;
+    }
+
+    // Rcout << std::endl;
+    // Rcout << std::endl;
+    // Rcout << std::endl;
+    //
+    // for(int i = 0; i<variables.size(); ++i){
+    //
+    //   // Rcout << variables[i].size();
+    //
+    //   for(auto dim: variables[i]) Rcout << dim << ",";
+    //
+    //   //  Rcout << variables[i][j] << ",";
+    //   //}
+    //
+    //   Rcout << std::endl;
+    // }
+
+    // ------------- purify -------------
+    // iterate backwards through tree family
+    int tree_index_t = curr_family.size() - 1;
+    for (auto tree_t = variables.rbegin(); tree_t != variables.rend(); ++tree_t)
+    {
+      std::set<int> curr_dims = *tree_t;
+      // do not purify null
+      if (curr_dims == std::set<int>{0})
+        continue;
+      // Rcout << std::endl << tree_index_t << " - T: ";
+      //  Rcout << "tree_t:";
+      //  for(auto dim: curr_dims) Rcout << dim << ", ";
+      //  Rcout << std::endl;
+
+      auto grid = grids[tree_index_t];
+      //     Rcout << "Grid dimensions of T: ";
+      //     for(auto dim: grid.dimensions) Rcout << dim << ", ";
+      //     Rcout << std::endl;
+      // go through subtrees of t
+      int tree_index_u = variables.size();
+      for (auto tree_u = variables.rbegin(); tree_u != variables.rend(); ++tree_u)
+      {
+        --tree_index_u;
+        // j_dims = dims of t without u
+        std::set<int> j_dims = curr_dims;
+        if (tree_u->size() > curr_dims.size())
+          continue;
+        // check if subset
+        bool subset = true;
+        for (const auto dim : *tree_u)
+        {
+          if (tree_t->count(dim) == 0)
+          {
+            subset = false;
+            break;
+          }
+          j_dims.erase(dim);
+        }
+        if (!subset)
+          continue;
+
+        // Rcout << "Hello";
+        // Rcout << "   " << tree_index_u << " - U: ";
+        // for(auto dim: *tree_u) Rcout << dim << ", ";
+        // Rcout << std::endl;
+        // Rcout << "   Individuals: ";
+
+        double tot_sum = 0;
+        grid = grids[tree_index_u];
+        while (!grid.nextPoint())
+        {
+          auto gridPoint = grid.getPoint();
+          //     Rcout << individuals[tree_index_u][gridPoint] << ", ";
+          tot_sum += individuals[tree_index_u][gridPoint];
+        }
+        // Rcout << "Total sum: " << tot_sum << std::endl;
+        // Rcout << std::endl;
+
+        //     Rcout << "      Grid dimensions of U: ";
+        //     for(auto dim: grid.dimensions) Rcout << dim << ", ";
+        //     Rcout << std::endl;
+
+        // Rcout<< "j_dims: "<<j_dims.size() << std::endl;;
+
+        std::vector<double> update(value_size, 0);
+
+        if (j_dims.size() == 0)
+        {
+
+          // grid = grids[tree_index_u];
+          while (!grid.nextPoint())
+          {
+            auto gridPoint_i = grid.getPoint();
+            //     Rcout << "         " << "i: ";
+            //     for(auto p: gridPoint_i) Rcout << p << ", ";
+            //     Rcout << std::endl << "         ";
+            double curr_sum = individuals[tree_index_u][gridPoint_i];
+            //     Rcout << ", Current Sum: " << curr_sum << std::endl;
+            //     Rcout << std::endl << "         " << "i, j: ";
+            update += (curr_sum / tot_sum) * values_old[tree_index_t][gridPoint_i];
+            //     Rcout << std::endl;
+          }
+
+          int tree_index_s = variables.size();
+          for (auto tree_s = variables.rbegin(); tree_s != variables.rend(); ++tree_s)
+          {
+
+            // Rcout << "tree_s:";
+            // for(auto dim: *tree_s) Rcout << dim << ", ";
+            // Rcout << std::endl;
+
+            --tree_index_s;
+            if (*tree_s == std::set<int>{0})
+            {
+
+              auto gridPoint_0 = std::vector<int>{0};
+              values[tree_index_s][gridPoint_0] += update;
+              //     Rcout << std::endl;
+              //}
+
+              /*
+               for(auto tree_0: curr_family){
+
+               if(tree_0.first == std::set<int>{0}){
+
+               Rcout << tree_0.first.size();
+               std::vector<int> leaf_index(tree_0.first.size(), 0);
+               std::vector<int> leaf_index(tree_0.second->GridLeaves.values.size(), 0);
+
+               int Test = tree_0.second->GridLeaves.values.size();
+               Rcout << Test;
+               tree_0.second->GridLeaves.values[leaf_index] += update;
+               }
+               }
+               */
+            }
+            else
+            {
+
+              // check if S subset of T
+
+              bool subset = true;
+              for (const auto dim : *tree_s)
+              {
+                if (tree_t->count(dim) == 0)
+                {
+                  subset = false;
+                  break;
+                }
+              }
+              if (!subset)
+                continue;
+
+              // Rcout << pow(-1, (*tree_s).size()) << std::endl;
+
+              auto grid_k = grids[tree_index_s];
+              while (!grid_k.nextPoint())
+              {
+                auto gridPoint_k = grid_k.getPoint();
+                //
+                //      if((*tree_s).size()>2){
+                //      Rcout << std::endl << "            " << "j, k: ";
+                //      for(auto p: gridPoint_k) Rcout << p << ", ";
+                //      Rcout << std::endl;
+                //      }
+                //
+                //      Rcout << pow(-1, (*tree_s).size()) * update << std::endl;
+                values[tree_index_s][gridPoint_k] += pow(-1, (*tree_s).size()) * update;
+              }
+            }
+          }
+          // Rcout << std::endl;
+        }
+        else
+        {
+
+          std::vector<int> j_sizes(j_dims.size(), 0);
+          for (size_t j = 0; j < j_dims.size(); ++j)
+          {
+            auto tmp = j_dims.begin();
+            std::advance(tmp, j);
+            int j_index = std::distance(variables[tree_index_t].begin(), variables[tree_index_t].find(*tmp));
+            j_sizes[j] = grids[tree_index_t].dimensions[j_index];
+          }
+
+          // Rcout<<"Hello 1";
+
+          grid::NDGrid grid_j = grid::NDGrid(j_sizes);
+          while (!grid_j.nextPoint()) // looping over every x_{i, T\U} essentially
+          {
+
+            std::vector<double> update(value_size, 0);
+            auto gridPoint_j = grid_j.getPoint();
+            //     Rcout << "         " << "j: ";
+            //     for(auto p: gridPoint_j) Rcout << p << ", ";
+            //     Rcout << std::endl;
+            // calc update
+            grid = grids[tree_index_u];
+            while (!grid.nextPoint())
+            {
+              auto gridPoint_i = grid.getPoint();
+              //     Rcout << "         " << "i: ";
+              //     for(auto p: gridPoint_i) Rcout << p << ", ";
+              //     Rcout << std::endl << "         ";
+              double curr_sum = individuals[tree_index_u][gridPoint_i];
+              //     Rcout << ", Current Sum: " << curr_sum << std::endl;
+              std::vector<int> gridPoint_ij(tree_t->size(), 0);
+              for (size_t j = 0; j < gridPoint_j.size(); ++j)
+              {
+                auto j_dim = j_dims.begin(); // to keep: T\U
+                std::advance(j_dim, j);
+                int j_index = std::distance(variables[tree_index_t].begin(), variables[tree_index_t].find(*j_dim));
+                //     Rcout << "         j_dim=" << *j_dim << ", j_index=" << j_index;
+                gridPoint_ij[j_index] = gridPoint_j[j];
+              }
+              for (size_t i = 0; i < gridPoint_i.size(); ++i)
+              {
+                auto i_dim = tree_u->begin(); // to marginalize: U
+                std::advance(i_dim, i);
+                int i_index = std::distance(variables[tree_index_t].begin(), variables[tree_index_t].find(*i_dim));
+                //     Rcout << "         i_dim=" << *i_dim << ", i_index=" << i_index;
+                gridPoint_ij[i_index] = gridPoint_i[i];
+              }
+              //     Rcout << std::endl << "         " << "i, j: ";
+              //     for(auto p: gridPoint_ij) Rcout << p << ", ";
+              //     Rcout << std::endl;
+              update += (curr_sum / tot_sum) * values_old[tree_index_t][gridPoint_ij];
+              //     Rcout << std::endl;
+            }
+
+            // Rcout << "Hello_2";
+            // update trees
+            int tree_index_s = variables.size();
+            for (auto tree_s = variables.rbegin(); tree_s != variables.rend(); ++tree_s)
+            {
+              --tree_index_s;
+              // check if T\U=j_dims subset of S and S subset of T
+              bool subset = true;
+              for (const auto dim : j_dims)
+              {
+                if (tree_s->count(dim) == 0)
+                {
+                  subset = false;
+                  break;
+                }
+              }
+              for (const auto dim : *tree_s)
+              {
+                if (tree_t->count(dim) == 0)
+                {
+                  subset = false;
+                  break;
+                }
+              }
+              if (!subset)
+                continue;
+              //     Rcout << "         " << "S: ";
+              //     for(auto dim: *tree_s) Rcout << dim << ", ";
+              //     Rcout << std::endl;
+              // S cap U
+              std::set<int> k_dims = *tree_s;
+              std::set<int> k_dims_h1 = *tree_s;
+              std::set<int> k_dims_h2 = *tree_u;
+              for (const auto dim : *tree_u)
+                k_dims.insert(dim);
+              for (const auto dim : *tree_s)
+                k_dims_h2.erase(dim);
+              for (const auto dim : *tree_u)
+                k_dims_h1.erase(dim);
+              for (const auto dim : k_dims_h1)
+                k_dims.erase(dim);
+              for (const auto dim : k_dims_h2)
+                k_dims.erase(dim);
+
+              // std::set<int> k_dims = *tree_s;
+              // for(const auto dim: *tree_t) k_dims.erase(dim);
+              // for(const auto dim: *tree_u) k_dims.insert(dim);
+
+              //     Rcout << "         " << "k_dims: ";
+              //     for(auto dim: k_dims) Rcout << dim << ", ";
+              //     Rcout << std::endl;
+
+              if (k_dims.size() == 0)
+              {
+
+                values[tree_index_s][gridPoint_j] += pow(-1, (*tree_s).size() - j_dims.size()) * update;
+              }
+              else
+              {
+
+                // Rcout <<"k_dims :";
+                // for(auto dim: k_dims) Rcout << dim << ", ";
+                // Rcout << std::endl;
+
+                std::vector<int> k_sizes(k_dims.size(), 0);
+                for (size_t k = 0; k < k_dims.size(); ++k)
+                {
+                  auto tmp = k_dims.begin();
+                  std::advance(tmp, k);
+                  int k_index = std::distance(variables[tree_index_t].begin(), variables[tree_index_t].find(*tmp));
+                  k_sizes[k] = grids[tree_index_t].dimensions[k_index];
+                }
+                // Rcout << "         " << "k_sizes: ";
+                // for(auto dim: k_sizes) Rcout << dim << ", ";
+                // Rcout << std::endl;
+                grid::NDGrid grid_k = grid::NDGrid(k_sizes);
+                while (!grid_k.nextPoint())
+                {
+                  auto gridPoint_k = grid_k.getPoint();
+                  // Rcout << "            " << "k: ";
+                  // for(auto p: gridPoint_k) Rcout << p << ", ";
+                  // Rcout << std::endl << "         ";
+                  std::vector<int> gridPoint_jk(tree_s->size(), 0);
+                  for (size_t j = 0; j < gridPoint_j.size(); ++j)
+                  {
+                    auto j_dim = j_dims.begin();
+                    std::advance(j_dim, j);
+                    int j_index = std::distance(variables[tree_index_s].begin(), variables[tree_index_s].find(*j_dim));
+                    // Rcout << "         j_dim=" << *j_dim << ", j_index=" << j_index;
+                    gridPoint_jk[j_index] = gridPoint_j[j];
+                  }
+                  for (size_t k = 0; k < gridPoint_k.size(); ++k)
+                  {
+                    auto k_dim = k_dims.begin();
+                    std::advance(k_dim, k);
+                    int k_index = std::distance(variables[tree_index_s].begin(), variables[tree_index_s].find(*k_dim));
+                    // Rcout << "         k_dim=" << *k_dim << ", k_index=" << k_index;
+                    gridPoint_jk[k_index] = gridPoint_k[k];
+                  }
+                  // Rcout << std::endl << "            " << "j, k: ";
+                  // for(auto p: gridPoint_jk) Rcout << p << ", ";
+                  // Rcout << std::endl;
+
+                  // Rcout << pow(-1, (*tree_s).size() - j_dims.size()) * update[0];
+                  values[tree_index_s][gridPoint_jk] += pow(-1, (*tree_s).size() - j_dims.size()) * update;
+                }
+              }
+            }
+          }
+        }
+      }
+      --tree_index_t;
+    }
+
+    // ------------- attach to rpf class -------------
+
+    // fill with new trees
+    for (size_t tree_index = 0; tree_index < variables.size(); ++tree_index)
+    {
+      LeafGrid curr_gridLeaf;
+      curr_gridLeaf.grid = grids[tree_index];
+      curr_gridLeaf.individuals = individuals[tree_index];
+      curr_gridLeaf.lim_list = lim_list;
+      curr_gridLeaf.values = values[tree_index];
+      curr_family[variables[tree_index]]->GridLeaves = curr_gridLeaf;
+    }
+  }
+
+  purified = true;
+}
+
+std::vector<std::vector<double>> RandomPlantedForest::get_lim_list(const TreeFamily &curr_family) {
+    // lim_list is a list giving for each variable all interval end-points
+    std::vector<std::vector<double>> lim_list(feature_size);
+
+    // go through all variables of the component
+    for (int curr_dim = 1; curr_dim <= feature_size; ++curr_dim)
+    {
+      std::vector<double> bounds;
+
+      // go through trees of family
+      for (const auto &curr_tree : curr_family)
+      {
+
+        // consider only relevant trees that have current dimension as variable
+        if (!curr_tree.first.count(curr_dim))
+          continue;
+
+        // go through leaves of tree
+        for (const auto &curr_leaf : curr_tree.second->leaves)
+        {
+          // get interval ends of variable
+          bounds.push_back(curr_leaf.intervals[curr_dim - 1].first);
+          bounds.push_back(curr_leaf.intervals[curr_dim - 1].second);
+        }
+      }
+      std::sort(bounds.begin(), bounds.end());
+      bounds.erase(std::unique(bounds.begin(), bounds.end()), bounds.end());
+      // int i_last = bounds.size()-1;
+      // double bibi = bounds[i_last] + 0.0001;
+      // bounds[i_last] = bounds[i_last] + 0.0001;
+      lim_list[curr_dim - 1] = bounds;
+    }
+
+    return lim_list;
+}
+
+
+void RandomPlantedForest::purify_1()
+{
+
+  // go through all n_trees families
+  for (auto &curr_family : this->tree_families)
+  {
+
+    // recap maximum number of dimensions of current family
+    unsigned int curr_max = 0;
+    for (auto tree : curr_family)
+    {
+      if (tree.first.size() > curr_max)
+        curr_max = tree.first.size();
+    }
+
+    while (curr_max >= 1)
+    {
+
+      // go through split dimensions of all trees
+      auto keys = getKeys(curr_family);
+      std::vector<std::set<int>>::reverse_iterator key = keys.rbegin();
+      while (key != keys.rend())
+      {
+
+        auto &curr_tree = curr_family[(*key)];
+        std::set<int> curr_dims = curr_tree->split_dims;
+
+        // check if number of dims same as current max_interaction
+        if (curr_dims.size() == curr_max)
+        {
+
+          // go through feature dims
+          for (int feature_dim = 1; feature_dim <= feature_size; ++feature_dim)
+          {
+
+            // continue only if dim in current tree
+            if (curr_tree->split_dims.count(feature_dim) != 0)
+            {
+
+              std::set<int> tree_dims = curr_tree->split_dims;
+              tree_dims.erase(tree_dims.find(feature_dim)); // remove current feature dim from current tree
+
+              // check if tree with dimensions exists, if not create
+              std::shared_ptr<DecisionTree> tree = treeExists(tree_dims, curr_family);
+              if (curr_max == 1)
+              {
+                tree = curr_family[std::set<int>{0}];
+              }
+              else
+              {
+                if (!tree)
+                {
+                  curr_family.insert(std::make_pair(tree_dims, std::make_shared<DecisionTree>(DecisionTree(tree_dims))));
+                  tree = curr_family[tree_dims];
+                }
+              }
+
+              // go through leaves of current tree
+              int n_leaves = curr_tree->leaves.size();
+              for (int l = 0; l < n_leaves; ++l)
+              {
+                auto &curr_leaf = curr_tree->leaves[l];
+
+                double multiplier = (curr_leaf.intervals[feature_dim - 1].second - curr_leaf.intervals[feature_dim - 1].first) / (upper_bounds[feature_dim - 1] - lower_bounds[feature_dim - 1]);
+
+                // new leaf including intervals and value
+                Leaf new_leaf = curr_leaf; // initialize intervals with first leaf
+                new_leaf.intervals[feature_dim - 1].first = lower_bounds[feature_dim - 1];
+                new_leaf.intervals[feature_dim - 1].second = upper_bounds[feature_dim - 1];
+                for (size_t i = 0; i < value_size; ++i)
+                  new_leaf.value[i] = -curr_leaf.value[i] * multiplier; // update value of new leaf
+
+                // append new leaf
+                if (!leafExists(new_leaf.intervals, curr_tree))
+                  curr_tree->leaves.push_back(new_leaf);
+                for (size_t i = 0; i < value_size; ++i)
+                  new_leaf.value[i] = curr_leaf.value[i] * multiplier; // update value of new leaf
+                if (!leafExists(new_leaf.intervals, tree))
+                  tree->leaves.push_back(new_leaf);
+              }
+            }
+          }
+        }
+        key++;
+      }
+
+      // update currently considered dimension size
+      --curr_max;
+    }
+  }
+
+  purified = true;
+}
+
+void RandomPlantedForest::purify_2()
+{
+
+  // go through all n_trees families
+  for (auto &curr_family : this->tree_families)
+  {
+
+    // lim_list is a list giving for each variable all interval end-points
+    std::vector<std::vector<double>> lim_list(feature_size);
+
+    // go through all variables of the component
+    for (int curr_dim = 1; curr_dim <= feature_size; ++curr_dim)
+    {
+      std::vector<double> bounds;
+
+      // go through trees of family
+      for (const auto &curr_tree : curr_family)
+      {
+
+        // consider only relevant trees that have current dimension as variable
+        if (!curr_tree.first.count(curr_dim))
+          continue;
+
+        // go through leaves of tree
+        for (const auto &curr_leaf : curr_tree.second->leaves)
+        {
+          // get interval ends of variable
+          bounds.push_back(curr_leaf.intervals[curr_dim - 1].second);
+        }
+      }
+      std::sort(bounds.begin(), bounds.end());
+      bounds.erase(std::unique(bounds.begin(), bounds.end()), bounds.end());
+      lim_list[curr_dim - 1] = bounds;
+    }
+
+    // initialize values and individuals for each tree in family
+    std::vector<grid::NDGrid> grids(curr_family.size() - 1);
+    std::vector<utils::Matrix<int>> individuals(curr_family.size() - 1);
+    std::vector<utils::Matrix<std::vector<double>>> values(curr_family.size() - 1);
+    std::vector<std::set<int>> variables(curr_family.size() - 1);
+
+    //  ------------- setup finer grid  -------------
+
+    int tree_index = 0;
+    for (const auto &curr_tree : curr_family)
+    {
+
+      if (curr_tree.first == std::set<int>{0})
+        continue; // ignore null tree
+
+      // fill space with dimensions
+      std::vector<int> dimensions;
+      for (const auto &dim : curr_tree.first)
+      {
+        dimensions.push_back(lim_list[dim - 1].size() - 1); // size - 1 ?
+      }
+
+      // setup grid for leaf indices
+      auto grid = grid::NDGrid(dimensions);
+
+      // initialize data for current tree
+      grids[tree_index] = grid;
+      individuals[tree_index] = utils::Matrix<int>(dimensions, 0);
+      values[tree_index] = utils::Matrix<std::vector<double>>(dimensions, std::vector<double>(value_size, 0)); // changed
+      variables[tree_index] = curr_tree.first;
+
+      // fill grid points with individuals and values
+      while (!grid.nextPoint())
+      {
+
+        std::vector<int> gridPoint = grid.getPoint();
+
+        bool in_leaf = true;
+
+        // go through sample points to sum up individuals
+        for (const auto &feature_vec : X)
+        {
+          int dim_index = 0;
+          in_leaf = true;
+          for (const auto &dim : curr_tree.first)
+          {
+            double val = feature_vec[dim - 1];
+            if (!((val >= lim_list[dim - 1][gridPoint[dim_index]]) && (val < lim_list[dim - 1][gridPoint[dim_index] + 1])))
+              in_leaf = false;
+            ++dim_index;
+          }
+
+          // consider individuals only if all in
+          if (in_leaf)
+            individuals[tree_index][gridPoint] += 1;
+        }
+
+        // go through leaves of tree to sum up values
+        for (const auto &leaf : curr_tree.second->get_leaves())
+        {
+
+          in_leaf = true;
+          int dim_index = 0;
+          for (const auto &dim : curr_tree.first)
+          {
+            // consider values only if all in
+            if (!((leaf.intervals[dim - 1].first <= lim_list[dim - 1][gridPoint[dim_index]]) && (leaf.intervals[dim - 1].second >= lim_list[dim - 1][gridPoint[dim_index] + 1])))
+              in_leaf = false;
+            ++dim_index;
+          }
+
+          // sum up values
+          if (in_leaf)
+            values[tree_index][gridPoint] += leaf.value; // todo: multiclass
+        }
+      }
+
+      ++tree_index;
+    }
+
+    // ------------- create new trees -------------
+
+    // insert null tree
+    grids.insert(grids.begin(), grid::NDGrid());
+    values.insert(values.begin(), utils::Matrix<std::vector<double>>(std::vector<int>{1}, std::vector<double>(value_size, 0)));
+    individuals.insert(individuals.begin(), utils::Matrix<int>(std::vector<int>{1}));
+    variables.insert(variables.begin(), std::set<int>{0});
+
+    // recap maximum number of dimensions of current family
+    unsigned int curr_max = 0;
+    for (const auto &tree : curr_family)
+    {
+      if (tree.first.size() > curr_max)
+        curr_max = tree.first.size();
+    }
+
+    auto keys = getKeys(curr_family);
+    while (curr_max > 1)
+    {
+
+      // go through split dimensions of all trees
+      for (std::vector<std::set<int>>::reverse_iterator key = keys.rbegin(); key != keys.rend(); ++key)
+      {
+
+        auto &curr_tree = curr_family[(*key)];
+        std::set<int> curr_dims = curr_tree->split_dims;
+
+        // check if number of dims same as current max_interaction
+        if (curr_dims.size() == curr_max)
+        {
+
+          // go through feature dims
+          int dim_index = 0;
+          for (int feature_dim = 1; feature_dim <= feature_size; ++feature_dim)
+          {
+
+            // continue only if dim in current tree
+            if (curr_tree->split_dims.count(feature_dim) != 0)
+            {
+
+              std::set<int> tree_dims = curr_tree->split_dims;
+              tree_dims.erase(tree_dims.find(feature_dim)); // remove current feature dim from current tree
+
+              // check if tree with dimensions exists, if not create
+              std::shared_ptr<DecisionTree> tree = treeExists(tree_dims, curr_family);
+              if (!tree)
+              {
+
+                // get index of old and new tree
+                auto old_tree_index = std::distance(std::begin(curr_family), curr_family.find(curr_tree->get_split_dims()));
+                curr_family.insert(std::make_pair(tree_dims, std::make_shared<DecisionTree>(DecisionTree(tree_dims))));
+                auto tree_index = std::distance(std::begin(curr_family), curr_family.find(tree_dims));
+
+                // remove matrix dimension of respective variable
+                std::vector<int> matrix_dimensions = values[old_tree_index].dims;
+                matrix_dimensions.erase(matrix_dimensions.begin() + dim_index);
+
+                // initialize data for new tree
+                auto grid = grid::NDGrid(matrix_dimensions);
+                grids.insert(grids.begin() + tree_index, grid);
+                values.insert(values.begin() + tree_index, utils::Matrix<std::vector<double>>(matrix_dimensions, std::vector<double>(0, value_size)));
+                individuals.insert(individuals.begin() + tree_index, utils::Matrix<int>(matrix_dimensions));
+                variables.insert(variables.begin() + tree_index, tree_dims);
+
+                // fill individuals of new trees
+                while (!grid.nextPoint())
+                {
+
+                  std::vector<int> gridPoint = grid.getPoint();
+                  bool in_leaf = true;
+
+                  // go through sample points to sum up individuals
+                  for (const auto &feature_vec : X)
+                  {
+                    int dim_index = 0;
+                    in_leaf = true;
+                    for (const auto &dim : tree_dims)
+                    {
+                      double val = feature_vec[dim - 1];
+                      if (!((val >= lim_list[dim - 1][gridPoint[dim_index]]) && (val < lim_list[dim - 1][gridPoint[dim_index] + 1])))
+                        in_leaf = false;
+                      ++dim_index;
+                    }
+
+                    // consider individuals only if all in
+                    if (in_leaf)
+                      individuals[tree_index][gridPoint] += 1;
+                  }
+                }
+              }
+
+              dim_index++;
+            }
+          }
+        }
+      }
+
+      // update currently considered dimension size
+      --curr_max;
+    }
+
+    // ------------- purify -------------
+
+    // measure tolerance and number of iterations
+    std::vector<double> tol(curr_family.size(), 1);
+    int iter;
+
+    // iterate backwards through tree family
+    int curr_tree_index = curr_family.size() - 1;
+    for (TreeFamily::reverse_iterator curr_tree = curr_family.rbegin(); curr_tree != curr_family.rend(); ++curr_tree)
+    {
+      iter = 0;
+      std::set<int> curr_dims = curr_tree->second->get_split_dims();
+
+      // do not purify null
+      if (curr_dims == std::set<int>{0})
+        continue;
+
+      // repeat until tolerance small enough and (?) maximum number of iterations reached
+      while ((tol[curr_tree_index] > 0.00000000001) && (iter < 100))
+      {
+
+        // go through feature dims
+        int curr_dim_index = 0;
+        for (const auto &feature_dim : curr_dims)
+        {
+
+          // get tree that has same variables as curr_tree minus j-variable
+          std::set<int> tree_dims = curr_dims;
+          tree_dims.erase(tree_dims.find(feature_dim));
+          int tree_index = 0; // if tree not exist, set to null tree
+          if (curr_family.find(tree_dims) != curr_family.end())
+            tree_index = std::distance(std::begin(curr_family), curr_family.find(tree_dims)) - 1;
+
+          // update values
+          if (grids[curr_tree_index].dimensions.size() == 1)
+          { // one dimensional case
+
+            int sum_ind = 0;
+            std::vector<double> avg(value_size, 0);
+
+            // get sum of individuals
+            for (int i = 0; i < individuals[curr_tree_index].n_entries; ++i)
+            {
+              std::vector<int> tmp{i};
+              sum_ind += individuals[curr_tree_index][tmp];
+            }
+            if (sum_ind == 0)
+              continue;
+
+            // calc avg
+            for (int i = 0; i < individuals[curr_tree_index].n_entries; ++i)
+            {
+              std::vector<int> tmp{i};
+              avg += (individuals[curr_tree_index][tmp] * values[curr_tree_index][tmp]) / sum_ind;
+            }
+
+            // update values of one dimensional and null tree
+            for (int i = 0; i < values[curr_tree_index].n_entries; ++i)
+            {
+              std::vector<int> tmp{i};
+              values[curr_tree_index][tmp] -= avg;
+            }
+            std::vector<int> tmp{0};
+            values[tree_index][tmp] += avg;
+          }
+          else
+          { // higher dimensional case
+
+            // setup new grid without dimension j
+            std::vector<int> new_dimensions = grids[curr_tree_index].dimensions;
+            int j_dim = new_dimensions[curr_dim_index];
+            new_dimensions.erase(new_dimensions.begin() + curr_dim_index);
+            grid::NDGrid grid = grid::NDGrid(new_dimensions);
+
+            // go through values without dimension j
+            while (!grid.nextPoint())
+            {
+              auto gridPoint = grid.getPoint();
+              gridPoint.push_back(0);
+
+              int sum_ind = 0;
+              std::vector<double> avg(value_size, 0);
+
+              // go through slice to sum up individuals
+              for (int j = 0; j < j_dim; ++j)
+              {
+                gridPoint.back() = j;
+
+                // get sum of individuals
+                sum_ind += individuals[curr_tree_index][gridPoint];
+              }
+
+              // go through slice to calc avg
+              for (int j = 0; j < j_dim; ++j)
+              {
+                gridPoint.back() = j;
+
+                // calc avg
+                avg += (individuals[curr_tree_index][gridPoint] * values[curr_tree_index][gridPoint]) / sum_ind;
+              }
+
+              // go through slice to update values
+              for (int j = 0; j < j_dim; ++j)
+              {
+                gridPoint.back() = j;
+
+                // update values of current slice
+                values[curr_tree_index][gridPoint] -= avg;
+              }
+
+              // update lower dimensional tree
+              gridPoint.pop_back();
+              values[tree_index][gridPoint] += avg;
+            }
+          }
+
+          ++curr_dim_index;
+        }
+
+        // update tolerance
+        if (variables[curr_tree_index].size() == 1)
+        {
+          tol[curr_tree_index] = 1; // todo
+        }
+        else
+        {
+          tol[curr_tree_index] = 1;
+        }
+
+        ++iter;
+      }
+
+      --curr_tree_index;
+    }
+
+    // ------------- attach to rpf class -------------
+
+    // fill with new trees
+    for (size_t tree_index = 0; tree_index < variables.size(); ++tree_index)
+    {
+      LeafGrid curr_gridLeaf;
+      curr_gridLeaf.grid = grids[tree_index];
+      curr_gridLeaf.individuals = individuals[tree_index];
+      curr_gridLeaf.lim_list = lim_list;
+      curr_gridLeaf.values = values[tree_index];
+      curr_family[variables[tree_index]]->GridLeaves = curr_gridLeaf;
+    }
+  }
+
+  purified = true;
+}
